refactor: move lessons type into types
