+++ src/components/CurriculumComponents/UnitModal/UnitModal.fixture.ts
@@ -1,5 +1,4 @@
-import { Lesson } from "@/components/CurriculumComponents/UnitModal/UnitModal";
-import { Unit } from "@/utils/curriculum/types";
+import { Unit, Lesson } from "@/utils/curriculum/types";
 
 const lessons: Lesson[] = [
   {
